+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Added:
 - field to set the country of origin for devices & platforms to support export control
 - checks to ensure that no duplicated contact role entries can be added
 - Tab to show the locations of all the configurations for a site
+- Pagination for device selection for tsm linkings
 
 Changed:
 - Devices can now also be mounted on other devices